fix getChatTitleOverride usage

+++ TMessagesProj/src/main/java/org/telegram/ui/PhotoViewer.java
@@ -11353,7 +11353,7 @@ public class PhotoViewer implements NotificationCenter.NotificationCenterDelegat
                                 chat = MessagesController.getInstance(currentAccount).getChat(newMessageObject.messageOwner.fwd_from.from_id.channel_id);
                             }
                             if (chat != null) {
-                                String title = UserConfig.getChatTitleOverride(currentAccount, (int)chat.id);
+                                String title = UserConfig.getChatTitleOverride(currentAccount, chat.id);
                                 if (title == null) {
                                     title = chat.title;
                                 }
@@ -11719,7 +11719,7 @@ public class PhotoViewer implements NotificationCenter.NotificationCenterDelegat
             if (parentChatActivity != null) {
                 TLRPC.Chat chat = parentChatActivity.getCurrentChat();
                 if (chat != null) {
-                    String title = UserConfig.getChatTitleOverride(currentAccount, (int)chat.id);
+                    String title = UserConfig.getChatTitleOverride(currentAccount, chat.id);
                     if (title == null) {
                         title = chat.title;
                     }
